refactor: split sync domains by concern
